Chat-Bot/0 conversational memory/requirements.txt
streamlit==1.24.0
PyPDF2==3.0.1
langchain==0.0.259
langchain_core==0.0.23
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.17.3
python-dotenv==1.0.0